requirements.txt
pillow
fastapi
uvicorn
gradio
requests
psutil
numpy
joblib
xgboost
category-encoders
scikit-learn